Register a new `gutenberg-for-node` package supporting Gutenberg library for node
