+++ bin/export-from-har.js
@@ -1,36 +1,6 @@
 require( 'fake-indexeddb/auto' );
-
-const noop = function () {};
-const { JSDOM, ResourceLoader } = require( 'jsdom' );
-class Window {
-	constructor( jsdomConfig = {} ) {
-		const { proxy, strictSSL, userAgent } = jsdomConfig;
-		const resources = new ResourceLoader( {
-			proxy,
-			strictSSL,
-			userAgent,
-		} );
-		return new JSDOM(
-			'',
-			Object.assign( jsdomConfig, {
-				resources,
-			} )
-		).window;
-	}
-}
-
-global.window = new Window( { url: 'http://localhost' } );
-global.document = window.document;
-global.requestAnimationFrame = global.cancelAnimationFrame = noop;
-global.navigator = window.navigator;
-global.Mousetrap = {
-	init: noop,
-	prototype: { stopCallback: noop },
-};
-window.matchMedia = global.matchMedia = () => ( { addListener: noop } );
-global.Node = window.Node;
-
-const { registerBlocks } = require( '../source/utils/register-blocks' );
+require( 'gutenberg-for-node' );
+const { registerBlocks } = require( 'site-parsers' ).utils;
 
 registerBlocks();
