+++ packages/site-parsers/src/utils/index.js
@@ -1,3 +1,4 @@
 module.exports = {
 	IdFactory: require( './idfactory' ),
+	...require( './register-blocks' ),
 };
