+++ packages/site-parsers/src/utils/register-blocks.js
